Added option to have the stabiliser seated instead of socketed
This stabiliser type should sit on top of the switchplate due to its height, or the keycap will not fully return to it's original position.
Socketed path will create a shape that the stabiliser can be inserted into, which is useful to create a stabiliser plate.
Seated path will cut out the shape required for the stabiliser to sit on the switch plate nicely, taking into account the connecting piece to the keycap requiring vertical space, as well as the stabiliser bar connection point requiring vertical space.
Technically the only required shape cutout is the seated cutout, but without a socketed cutout there is nothing holding the stabiliser in place other than adhesive.